refactor with new tests, ModelMeta in abstract
