+++ datamodel/__init__.py
@@ -3,13 +3,10 @@
 
 DataModel is a reimplementation of dataclasses with true inheritance and composition.
 """
-from typing import Optional, Union
-from collections.abc import Callable
-
 from datamodel.fields import Field, Column, fields
-from .base import BaseModel
+from .base import Model, BaseModel
 from .version import (
     __title__, __description__, __version__, __author__, __author_email__
 )
 
-# __all__ = ('fields', 'Field', 'Column', 'BaseModel', )
+__all__ = ('fields', 'Field', 'Column', 'Model', 'BaseModel', )
